post method in database running perfect
